Simplify service accounts and pod security policies
Includes bump to fissile version that implements the simplified model.

Jobs no longer specify a psp; there is just one psp anymore and it is
bound to all service accounts.  In addition to capabilities each job
now also specifies if the pod needs to be privileged.

Most of the cluster roles have been replaced by regular roles. The
exceptions are the eirini, node-reader, and test cluster roles.

Eirini continues to run effectively as cluster-admin and will be
addressed at a later time. The node-reader cluster role is required
for diego-cells to be able to read the AZ label of the nodes they
are running on. The test-role is used by brain tests and not installed
by the helm chart.

See also: https://github.com/cloudfoundry-incubator/fissile/wiki/ServiceAccount-and-PodSecurityPolicy-usage